requirements-railway.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Jinja2==3.1.2
SQLAlchemy==2.0.20
Werkzeug==2.3.7
WTForms==3.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
alembic==1.15.1
email-validator==2.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
numpy>=1.20.0
pandas==1.5.3
openpyxl==3.1.2
Pillow==10.0.0
setuptools>=65.5.1
wheel>=0.38.0
